feat: refactor Router
-add Component Handler
